lmkd: fix over-aggressive kills on direct reclaim
A total havoc was observed with a simple command like
`cat /dev/block/sd* > /dev/null`. lmkd starts killing *all* apps
within a few seconds after free memory for the page cache has been
depleted, including the foreground ones.

The worst part is that this behavior can be triggered from a
non-superuser, regular app. This is unacceptable, and must be fixed.

After a few debugging sessions, it turns out that the
wmark <= WMARK_HIGH check is too weak. Direct reclaim scenarios are
checked together with page cache thrasing from the logic right above
this, so this check is bogus (and obviously broken).

Fixes: f56cfa948de9 ("Improve ULMK performance.")
Change-Id: Ia25edb6104a77dfed700c939fa95a231e97eb285
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: rk134 <me@rk134.cf>